Removed duplicate components
