+++ finale/src/components/state.jsx
@@ -1,6 +1,8 @@
 import React, { Component } from 'react';
 import District from './district';
 import data from "../data"
+import dl_logo from "../img/download_icon.png"
+import state_logo from "../img/State_logo.png"
 
 class State extends Component {
   constructor() {
@@ -13,11 +15,17 @@ class State extends Component {
   //Click handler for expanding rows
   handleRowClick(rowId) {
     const currentExpandedRows = this.state.expandedRows;
+    console.log(currentExpandedRows);
+
+    //If false that means that there is no district expanded row at the time of this click handler
     const isRowCurrentlyExpanded = currentExpandedRows.includes(rowId);
 
+    //If isRowCurrentlyExpanded is false state row will expand
     const newExpandedRows = isRowCurrentlyExpanded ?
       currentExpandedRows.filter(id => id !== rowId) :
       currentExpandedRows.concat(rowId);
+
+    console.log(newExpandedRows);
 
     this.setState({ expandedRows: newExpandedRows });
   }
@@ -25,10 +33,21 @@ class State extends Component {
   //Rendering function for district row
   renderItem(sta) {
     const clickCallback = () => this.handleRowClick(sta.id);
+
     const stateRow = [
-      <tr className="names" key={"row-data-" + sta.id}>
-        <td> {sta.title}
-          <button className="toggle-btn" onClick={clickCallback}>{sta.subRegions.length} Districts</button>
+      <tr className="names" key={"state-row-data-" + sta.id}>
+        <td>
+          <img className="dl_logo" src={state_logo} alt="state_logo" />
+          <span>{sta.title}</span>
+          <img className="dl_logo" src={dl_logo} alt="dl_icon" />
+
+          {sta.subRegions.length > 0 &&
+            <button
+              className="toggle-btn"
+              onClick={clickCallback}>{sta.subRegions.length} District
+              {this.state.expandedRows > 0 ? " -" : " +"}
+            </button>
+          }
         </td>
         <td>{sta.lastIn}</td>
         <td>{sta.numForms}</td>
@@ -40,17 +59,8 @@ class State extends Component {
     //Toggle for rendering township rows of the district if there are townships under the district
     if (sta.subRegions.length > 0 && this.state.expandedRows.includes(sta.id)) {
       stateRow.push(
-
-        // <District   />
-
         sta.subRegions.map((district) =>
-          <tr className="names" key={"row-expanded-" + district.id}>
-            <td>{district.title}</td>
-            <td>{district.lastIn}</td>
-            <td>{district.numForms}</td>
-            <td>{district.numVotes}</td>
-            <td>{district.update}</td>
-          </tr>
+          <District district={district} />
         )
       )
     }
@@ -66,7 +76,7 @@ class State extends Component {
     });
 
     return (
-      <tr>{allItemRows}</tr>
+      <table className="names">{allItemRows}</table>
     );
   }
 }
